test(core): split finite-y-cache characterization by stat path (#503)
Move the 617-line finite-y-cache monolith into finite-y-cache/:
smooth-path, summary-boxplot, count-buckets, facet-local (+ fixtures).
Assertions unchanged (10 cases).

Closes #503.
