+++ packages/core/tests/candidate-construction/finite-y-cache/fixtures.ts
@@ -1,0 +1,4 @@
+/**
+ * Shared helpers for finite-y lineage cache characterization (issue #216 / #503).
+ */
+export { size } from "../fixtures.ts";
